+++ handlers/influx-db/main_test.go
@@ -1,0 +1,68 @@
+package main
+
+import (
+	"encoding/json"
+	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
+	"os"
+	"testing"
+
+	"github.com/sensu/sensu-go/types"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestSendMetrics(t *testing.T) {
+	assert := assert.New(t)
+	event := types.FixtureEvent("entity1", "check1")
+	event.Check = nil
+	event.Metrics = types.FixtureMetrics()
+
+	var apiStub = httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		body, _ := ioutil.ReadAll(r.Body)
+		expectedBody := `answer,foo=bar value=42`
+		assert.Contains(string(body), expectedBody)
+		w.WriteHeader(http.StatusOK)
+		_, err := w.Write([]byte(`{"ok": true}`))
+		require.NoError(t, err)
+	}))
+
+	addr = apiStub.URL
+	err := sendMetrics(event)
+	assert.NoError(err)
+}
+
+func TestMain(t *testing.T) {
+	assert := assert.New(t)
+	file, _ := ioutil.TempFile(os.TempDir(), "sensu-handler-influx-db-")
+	defer func() {
+		_ = os.Remove(file.Name())
+	}()
+
+	event := types.FixtureEvent("entity1", "check1")
+	event.Check = nil
+	event.Metrics = types.FixtureMetrics()
+	eventJSON, _ := json.Marshal(event)
+	_, err := file.WriteString(string(eventJSON))
+	require.NoError(t, err)
+	require.NoError(t, file.Sync())
+	_, err = file.Seek(0, 0)
+	require.NoError(t, err)
+	stdin = file
+	requestReceived := false
+
+	var apiStub = httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		requestReceived = true
+		w.WriteHeader(http.StatusOK)
+		_, err := w.Write([]byte(`{"ok": true}`))
+		require.NoError(t, err)
+	}))
+
+	oldArgs := os.Args
+	os.Args = []string{"influx-db", "-a", apiStub.URL, "-d", "foo", "-u", "bar", "-p", "baz"}
+	defer func() { os.Args = oldArgs }()
+
+	main()
+	assert.True(requestReceived)
+}