+++ CHANGELOG.md
@@ -35,6 +35,7 @@ event handlers for statsd metrics.
 - Add the `--statsd-disable` flag to sensu-agent which configures the
 statsd listener. The listener is enabled by default.
 - Add 'remove-when' and 'set-when' subcommands to sensuctl filter command.
+- Added an influx-db handler for events containing metrics.
 
 ### Changed
 - Changed the maximum number of open file descriptors on a system to from 1024